added error message to http request
