+++ src/app/app.module.ts
@@ -3,8 +3,9 @@ import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { HttpErrorHandler }     from './error-handler.service';
+import { HttpErrorHandler }     from './http-error-handler.service';
 import { MessageService }       from './message.service';
+import { MessagesComponent } from './message/message.component'
 import { ProfileImageComponent } from './profile-image/profile-image.component';
 import { HeaderComponent } from './common/header.component';
 import { FooterComponent } from './common/footer.component';
@@ -19,7 +20,8 @@ import { ProfileSearchComponent } from './profile-search/profile-search.componen
     FooterComponent,
     ProfileImageComponent,
     ShowProfileComponent,
-    ProfileSearchComponent
+    ProfileSearchComponent,
+    MessagesComponent
   ],
   imports: [
     BrowserModule,
