--- src/app/in-memory-data.service.ts
@@ -1,1420 +1,0 @@
-import { Injectable } from '@angular/core';
-import { InMemoryDbService } from 'angular-in-memory-web-api';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class InMemoryDataService implements InMemoryDbService {
-  createDb() {
-    const repoInfo = [
-      {
-        "id": 137380096,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMzczODAwOTY=",
-        "name": "AngularGithubProfile",
-        "full_name": "DCox2016/AngularGithubProfile",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/AngularGithubProfile",
-        "description": null,
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/AngularGithubProfile",
-        "forks_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/AngularGithubProfile/deployments",
-        "created_at": "2018-06-14T15:59:33Z",
-        "updated_at": "2018-06-14T15:59:33Z",
-        "pushed_at": "2018-06-14T15:59:33Z",
-        "git_url": "git://github.com/DCox2016/AngularGithubProfile.git",
-        "ssh_url": "git@github.com:DCox2016/AngularGithubProfile.git",
-        "clone_url": "https://github.com/DCox2016/AngularGithubProfile.git",
-        "svn_url": "https://github.com/DCox2016/AngularGithubProfile",
-        "homepage": null,
-        "size": 0,
-        "stargazers_count": 3,
-        "watchers_count": 0,
-        "language": "angular",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 133973107,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMzM5NzMxMDc=",
-        "name": "AngularJScode",
-        "full_name": "DCox2016/AngularJScode",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/AngularJScode",
-        "description": null,
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/AngularJScode",
-        "forks_url": "https://api.github.com/repos/DCox2016/AngularJScode/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/AngularJScode/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/AngularJScode/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/AngularJScode/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/AngularJScode/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/AngularJScode/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/AngularJScode/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/AngularJScode/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/AngularJScode/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/AngularJScode/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/AngularJScode/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/AngularJScode/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/AngularJScode/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/AngularJScode/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/AngularJScode/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/AngularJScode/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/AngularJScode/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/AngularJScode/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/AngularJScode/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/AngularJScode/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/AngularJScode/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/AngularJScode/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/AngularJScode/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/AngularJScode/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/AngularJScode/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/AngularJScode/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/AngularJScode/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/AngularJScode/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/AngularJScode/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/AngularJScode/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/AngularJScode/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/AngularJScode/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/AngularJScode/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/AngularJScode/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/AngularJScode/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/AngularJScode/deployments",
-        "created_at": "2018-05-18T15:34:48Z",
-        "updated_at": "2018-09-13T17:51:04Z",
-        "pushed_at": "2018-05-20T20:06:47Z",
-        "git_url": "git://github.com/DCox2016/AngularJScode.git",
-        "ssh_url": "git@github.com:DCox2016/AngularJScode.git",
-        "clone_url": "https://github.com/DCox2016/AngularJScode.git",
-        "svn_url": "https://github.com/DCox2016/AngularJScode",
-        "homepage": null,
-        "size": 4,
-        "stargazers_count": 1,
-        "watchers_count": 1,
-        "language": "HTML",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 1,
-        "default_branch": "master"
-      },
-      {
-        "id": 137380158,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMzczODAxNTg=",
-        "name": "Angularv6GithubProfile",
-        "full_name": "DCox2016/Angularv6GithubProfile",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Angularv6GithubProfile",
-        "description": "Uses Angular v6 to search Github and display profile information",
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile",
-        "forks_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Angularv6GithubProfile/deployments",
-        "created_at": "2018-06-14T16:00:18Z",
-        "updated_at": "2018-06-27T11:55:44Z",
-        "pushed_at": "2018-06-27T11:56:24Z",
-        "git_url": "git://github.com/DCox2016/Angularv6GithubProfile.git",
-        "ssh_url": "git@github.com:DCox2016/Angularv6GithubProfile.git",
-        "clone_url": "https://github.com/DCox2016/Angularv6GithubProfile.git",
-        "svn_url": "https://github.com/DCox2016/Angularv6GithubProfile",
-        "homepage": null,
-        "size": 100,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "TypeScript",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 135187585,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMzUxODc1ODU=",
-        "name": "Angularv6Heros",
-        "full_name": "DCox2016/Angularv6Heros",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Angularv6Heros",
-        "description": null,
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Angularv6Heros",
-        "forks_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Angularv6Heros/deployments",
-        "created_at": "2018-05-28T16:55:23Z",
-        "updated_at": "2018-05-31T13:19:10Z",
-        "pushed_at": "2018-05-31T13:19:08Z",
-        "git_url": "git://github.com/DCox2016/Angularv6Heros.git",
-        "ssh_url": "git@github.com:DCox2016/Angularv6Heros.git",
-        "clone_url": "https://github.com/DCox2016/Angularv6Heros.git",
-        "svn_url": "https://github.com/DCox2016/Angularv6Heros",
-        "homepage": null,
-        "size": 187,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "TypeScript",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 162313584,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNjIzMTM1ODQ=",
-        "name": "Dustinsdomain",
-        "full_name": "DCox2016/Dustinsdomain",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Dustinsdomain",
-        "description": "This will be my homepage with information about me.",
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Dustinsdomain",
-        "forks_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Dustinsdomain/deployments",
-        "created_at": "2018-12-18T16:07:28Z",
-        "updated_at": "2018-12-18T21:11:52Z",
-        "pushed_at": "2018-12-18T21:11:50Z",
-        "git_url": "git://github.com/DCox2016/Dustinsdomain.git",
-        "ssh_url": "git@github.com:DCox2016/Dustinsdomain.git",
-        "clone_url": "https://github.com/DCox2016/Dustinsdomain.git",
-        "svn_url": "https://github.com/DCox2016/Dustinsdomain",
-        "homepage": null,
-        "size": 45354,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 116269782,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMTYyNjk3ODI=",
-        "name": "Git-test",
-        "full_name": "DCox2016/Git-test",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Git-test",
-        "description": "Node.JS training",
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Git-test",
-        "forks_url": "https://api.github.com/repos/DCox2016/Git-test/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Git-test/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Git-test/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Git-test/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Git-test/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Git-test/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Git-test/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Git-test/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Git-test/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Git-test/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Git-test/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Git-test/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Git-test/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Git-test/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Git-test/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Git-test/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Git-test/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Git-test/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Git-test/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Git-test/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Git-test/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Git-test/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Git-test/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Git-test/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Git-test/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Git-test/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Git-test/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Git-test/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Git-test/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Git-test/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Git-test/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Git-test/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Git-test/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Git-test/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Git-test/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Git-test/deployments",
-        "created_at": "2018-01-04T14:33:36Z",
-        "updated_at": "2018-01-04T15:39:56Z",
-        "pushed_at": "2018-01-04T15:39:55Z",
-        "git_url": "git://github.com/DCox2016/Git-test.git",
-        "ssh_url": "git@github.com:DCox2016/Git-test.git",
-        "clone_url": "https://github.com/DCox2016/Git-test.git",
-        "svn_url": "https://github.com/DCox2016/Git-test",
-        "homepage": null,
-        "size": 0,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 151611259,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTE2MTEyNTk=",
-        "name": "GithubProfileViewerwithAngularv6",
-        "full_name": "DCox2016/GithubProfileViewerwithAngularv6",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/GithubProfileViewerwithAngularv6",
-        "description": null,
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6",
-        "forks_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/GithubProfileViewerwithAngularv6/deployments",
-        "created_at": "2018-10-04T17:36:13Z",
-        "updated_at": "2018-11-30T08:22:11Z",
-        "pushed_at": "2018-11-30T08:22:09Z",
-        "git_url": "git://github.com/DCox2016/GithubProfileViewerwithAngularv6.git",
-        "ssh_url": "git@github.com:DCox2016/GithubProfileViewerwithAngularv6.git",
-        "clone_url": "https://github.com/DCox2016/GithubProfileViewerwithAngularv6.git",
-        "svn_url": "https://github.com/DCox2016/GithubProfileViewerwithAngularv6",
-        "homepage": null,
-        "size": 45482,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "TypeScript",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 151558043,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTE1NTgwNDM=",
-        "name": "hacktoberfest",
-        "full_name": "DCox2016/hacktoberfest",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/hacktoberfest",
-        "description": "Hacktoberfest 2018",
-        "fork": true,
-        "url": "https://api.github.com/repos/DCox2016/hacktoberfest",
-        "forks_url": "https://api.github.com/repos/DCox2016/hacktoberfest/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/hacktoberfest/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/hacktoberfest/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/hacktoberfest/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/hacktoberfest/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/hacktoberfest/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/hacktoberfest/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/hacktoberfest/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/hacktoberfest/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/hacktoberfest/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/hacktoberfest/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/hacktoberfest/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/hacktoberfest/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/hacktoberfest/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/hacktoberfest/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/hacktoberfest/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/hacktoberfest/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/hacktoberfest/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/hacktoberfest/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/hacktoberfest/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/hacktoberfest/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/hacktoberfest/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/hacktoberfest/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/hacktoberfest/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/hacktoberfest/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/hacktoberfest/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/hacktoberfest/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/hacktoberfest/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/hacktoberfest/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/hacktoberfest/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/hacktoberfest/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/hacktoberfest/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/hacktoberfest/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/hacktoberfest/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/hacktoberfest/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/hacktoberfest/deployments",
-        "created_at": "2018-10-04T10:59:36Z",
-        "updated_at": "2018-10-04T15:11:40Z",
-        "pushed_at": "2018-10-04T15:11:39Z",
-        "git_url": "git://github.com/DCox2016/hacktoberfest.git",
-        "ssh_url": "git@github.com:DCox2016/hacktoberfest.git",
-        "clone_url": "https://github.com/DCox2016/hacktoberfest.git",
-        "svn_url": "https://github.com/DCox2016/hacktoberfest",
-        "homepage": "https://hacktoberfest.lingonsaft.com/",
-        "size": 7197,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": false,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": {
-          "key": "mit",
-          "name": "MIT License",
-          "spdx_id": "MIT",
-          "url": "https://api.github.com/licenses/mit",
-          "node_id": "MDc6TGljZW5zZTEz"
-        },
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 151561083,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTE1NjEwODM=",
-        "name": "Hacktoberfest-Census",
-        "full_name": "DCox2016/Hacktoberfest-Census",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Hacktoberfest-Census",
-        "description": "A census of those participating in Hacktoberfest (and an easy PR!)",
-        "fork": true,
-        "url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census",
-        "forks_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Hacktoberfest-Census/deployments",
-        "created_at": "2018-10-04T11:27:34Z",
-        "updated_at": "2018-10-04T11:36:30Z",
-        "pushed_at": "2018-10-04T11:36:28Z",
-        "git_url": "git://github.com/DCox2016/Hacktoberfest-Census.git",
-        "ssh_url": "git@github.com:DCox2016/Hacktoberfest-Census.git",
-        "clone_url": "https://github.com/DCox2016/Hacktoberfest-Census.git",
-        "svn_url": "https://github.com/DCox2016/Hacktoberfest-Census",
-        "homepage": "https://cutwell.github.io/Hacktoberfest-Census/",
-        "size": 688,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": false,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 151567534,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTE1Njc1MzQ=",
-        "name": "Hacktoberfest2018",
-        "full_name": "DCox2016/Hacktoberfest2018",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Hacktoberfest2018",
-        "description": "Welcome! I need some help making my list... Learn how in the README",
-        "fork": true,
-        "url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018",
-        "forks_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Hacktoberfest2018/deployments",
-        "created_at": "2018-10-04T12:23:35Z",
-        "updated_at": "2018-10-04T12:23:37Z",
-        "pushed_at": "2018-10-04T14:50:06Z",
-        "git_url": "git://github.com/DCox2016/Hacktoberfest2018.git",
-        "ssh_url": "git@github.com:DCox2016/Hacktoberfest2018.git",
-        "clone_url": "https://github.com/DCox2016/Hacktoberfest2018.git",
-        "svn_url": "https://github.com/DCox2016/Hacktoberfest2018",
-        "homepage": "https://raidandfade.github.io/Hacktoberfest2018",
-        "size": 284,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": false,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": false,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": {
-          "key": "mit",
-          "name": "MIT License",
-          "spdx_id": "MIT",
-          "url": "https://api.github.com/licenses/mit",
-          "node_id": "MDc6TGljZW5zZTEz"
-        },
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 101998849,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMDE5OTg4NDk=",
-        "name": "Hide-gnv-seek",
-        "full_name": "DCox2016/Hide-gnv-seek",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Hide-gnv-seek",
-        "description": "for css practice ",
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek",
-        "forks_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Hide-gnv-seek/deployments",
-        "created_at": "2017-08-31T12:34:04Z",
-        "updated_at": "2017-08-31T12:34:04Z",
-        "pushed_at": "2017-08-31T12:34:04Z",
-        "git_url": "git://github.com/DCox2016/Hide-gnv-seek.git",
-        "ssh_url": "git@github.com:DCox2016/Hide-gnv-seek.git",
-        "clone_url": "https://github.com/DCox2016/Hide-gnv-seek.git",
-        "svn_url": "https://github.com/DCox2016/Hide-gnv-seek",
-        "homepage": null,
-        "size": 0,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": null,
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 139057621,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxMzkwNTc2MjE=",
-        "name": "httpPlaywithAngular",
-        "full_name": "DCox2016/httpPlaywithAngular",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/httpPlaywithAngular",
-        "description": null,
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular",
-        "forks_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/httpPlaywithAngular/deployments",
-        "created_at": "2018-06-28T19:00:05Z",
-        "updated_at": "2018-06-28T19:01:09Z",
-        "pushed_at": "2018-06-28T19:01:08Z",
-        "git_url": "git://github.com/DCox2016/httpPlaywithAngular.git",
-        "ssh_url": "git@github.com:DCox2016/httpPlaywithAngular.git",
-        "clone_url": "https://github.com/DCox2016/httpPlaywithAngular.git",
-        "svn_url": "https://github.com/DCox2016/httpPlaywithAngular",
-        "homepage": null,
-        "size": 101,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "TypeScript",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 158268639,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTgyNjg2Mzk=",
-        "name": "Java-30",
-        "full_name": "DCox2016/Java-30",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Java-30",
-        "description": "My Java 30 solutions",
-        "fork": false,
-        "url": "https://api.github.com/repos/DCox2016/Java-30",
-        "forks_url": "https://api.github.com/repos/DCox2016/Java-30/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Java-30/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Java-30/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Java-30/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Java-30/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Java-30/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Java-30/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Java-30/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Java-30/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Java-30/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Java-30/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Java-30/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Java-30/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Java-30/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Java-30/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Java-30/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Java-30/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Java-30/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Java-30/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Java-30/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Java-30/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Java-30/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Java-30/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Java-30/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Java-30/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Java-30/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Java-30/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Java-30/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Java-30/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Java-30/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Java-30/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Java-30/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Java-30/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Java-30/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Java-30/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Java-30/deployments",
-        "created_at": "2018-11-19T17:59:28Z",
-        "updated_at": "2018-12-17T16:41:13Z",
-        "pushed_at": "2018-12-17T16:41:22Z",
-        "git_url": "git://github.com/DCox2016/Java-30.git",
-        "ssh_url": "git@github.com:DCox2016/Java-30.git",
-        "clone_url": "https://github.com/DCox2016/Java-30.git",
-        "svn_url": "https://github.com/DCox2016/Java-30",
-        "homepage": null,
-        "size": 54880,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": true,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 158255194,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTgyNTUxOTQ=",
-        "name": "JavaScript30",
-        "full_name": "DCox2016/JavaScript30",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/JavaScript30",
-        "description": "30 Day Vanilla JS Challenge",
-        "fork": true,
-        "url": "https://api.github.com/repos/DCox2016/JavaScript30",
-        "forks_url": "https://api.github.com/repos/DCox2016/JavaScript30/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/JavaScript30/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/JavaScript30/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/JavaScript30/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/JavaScript30/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/JavaScript30/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/JavaScript30/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/JavaScript30/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/JavaScript30/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/JavaScript30/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/JavaScript30/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/JavaScript30/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/JavaScript30/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/JavaScript30/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/JavaScript30/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/JavaScript30/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/JavaScript30/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/JavaScript30/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/JavaScript30/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/JavaScript30/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/JavaScript30/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/JavaScript30/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/JavaScript30/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/JavaScript30/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/JavaScript30/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/JavaScript30/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/JavaScript30/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/JavaScript30/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/JavaScript30/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/JavaScript30/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/JavaScript30/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/JavaScript30/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/JavaScript30/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/JavaScript30/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/JavaScript30/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/JavaScript30/deployments",
-        "created_at": "2018-11-19T16:17:40Z",
-        "updated_at": "2018-11-19T16:17:42Z",
-        "pushed_at": "2018-11-16T03:41:12Z",
-        "git_url": "git://github.com/DCox2016/JavaScript30.git",
-        "ssh_url": "git@github.com:DCox2016/JavaScript30.git",
-        "clone_url": "https://github.com/DCox2016/JavaScript30.git",
-        "svn_url": "https://github.com/DCox2016/JavaScript30",
-        "homepage": "https://JavaScript30.com",
-        "size": 45718,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": "HTML",
-        "has_issues": false,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      },
-      {
-        "id": 151562513,
-        "node_id": "MDEwOlJlcG9zaXRvcnkxNTE1NjI1MTM=",
-        "name": "Make-a-Pull-Request",
-        "full_name": "DCox2016/Make-a-Pull-Request",
-        "private": false,
-        "owner": {
-          "login": "DCox2016",
-          "id": 17097992,
-          "node_id": "MDQ6VXNlcjE3MDk3OTky",
-          "avatar_url": "https://avatars1.githubusercontent.com/u/17097992?v=4",
-          "gravatar_id": "",
-          "url": "https://api.github.com/users/DCox2016",
-          "html_url": "https://github.com/DCox2016",
-          "followers_url": "https://api.github.com/users/DCox2016/followers",
-          "following_url": "https://api.github.com/users/DCox2016/following{/other_user}",
-          "gists_url": "https://api.github.com/users/DCox2016/gists{/gist_id}",
-          "starred_url": "https://api.github.com/users/DCox2016/starred{/owner}{/repo}",
-          "subscriptions_url": "https://api.github.com/users/DCox2016/subscriptions",
-          "organizations_url": "https://api.github.com/users/DCox2016/orgs",
-          "repos_url": "https://api.github.com/users/DCox2016/repos",
-          "events_url": "https://api.github.com/users/DCox2016/events{/privacy}",
-          "received_events_url": "https://api.github.com/users/DCox2016/received_events",
-          "type": "User",
-          "site_admin": false
-        },
-        "html_url": "https://github.com/DCox2016/Make-a-Pull-Request",
-        "description": "Make a Pull Request",
-        "fork": true,
-        "url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request",
-        "forks_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/forks",
-        "keys_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/keys{/key_id}",
-        "collaborators_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/collaborators{/collaborator}",
-        "teams_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/teams",
-        "hooks_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/hooks",
-        "issue_events_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/issues/events{/number}",
-        "events_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/events",
-        "assignees_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/assignees{/user}",
-        "branches_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/branches{/branch}",
-        "tags_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/tags",
-        "blobs_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/git/blobs{/sha}",
-        "git_tags_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/git/tags{/sha}",
-        "git_refs_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/git/refs{/sha}",
-        "trees_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/git/trees{/sha}",
-        "statuses_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/statuses/{sha}",
-        "languages_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/languages",
-        "stargazers_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/stargazers",
-        "contributors_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/contributors",
-        "subscribers_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/subscribers",
-        "subscription_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/subscription",
-        "commits_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/commits{/sha}",
-        "git_commits_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/git/commits{/sha}",
-        "comments_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/comments{/number}",
-        "issue_comment_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/issues/comments{/number}",
-        "contents_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/contents/{+path}",
-        "compare_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/compare/{base}...{head}",
-        "merges_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/merges",
-        "archive_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/{archive_format}{/ref}",
-        "downloads_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/downloads",
-        "issues_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/issues{/number}",
-        "pulls_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/pulls{/number}",
-        "milestones_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/milestones{/number}",
-        "notifications_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/notifications{?since,all,participating}",
-        "labels_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/labels{/name}",
-        "releases_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/releases{/id}",
-        "deployments_url": "https://api.github.com/repos/DCox2016/Make-a-Pull-Request/deployments",
-        "created_at": "2018-10-04T11:40:40Z",
-        "updated_at": "2018-10-04T11:40:42Z",
-        "pushed_at": "2018-10-04T12:17:08Z",
-        "git_url": "git://github.com/DCox2016/Make-a-Pull-Request.git",
-        "ssh_url": "git@github.com:DCox2016/Make-a-Pull-Request.git",
-        "clone_url": "https://github.com/DCox2016/Make-a-Pull-Request.git",
-        "svn_url": "https://github.com/DCox2016/Make-a-Pull-Request",
-        "homepage": "",
-        "size": 326,
-        "stargazers_count": 0,
-        "watchers_count": 0,
-        "language": null,
-        "has_issues": false,
-        "has_projects": true,
-        "has_downloads": true,
-        "has_wiki": true,
-        "has_pages": false,
-        "forks_count": 0,
-        "mirror_url": null,
-        "archived": false,
-        "open_issues_count": 0,
-        "license": null,
-        "forks": 0,
-        "open_issues": 0,
-        "watchers": 0,
-        "default_branch": "master"
-      }
-    ];
-    return {repoInfo};
-    }
-}
