+++ src/app/message/message.component.html
@@ -1,0 +1,5 @@
+<div *ngIf="messageService.messages.length">
+    <h3>Error Message:</h3>
+    <p *ngFor='let message of messageService.messages'> {{message}} </p>
+    <button class="clear" (click)="messageService.clear()">clear</button>
+  </div>
