+++ src/app/message/message.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { MessageService } from '../message.service';
+
+@Component({
+  selector: 'ghv-app-messages',
+  templateUrl: './message.component.html'
+})
+export class MessagesComponent {
+  constructor(public messageService: MessageService) {}
+}
