+++ src/app/profile-search/profile-search.component.ts
@@ -9,6 +9,7 @@ import { ProfileSearchService  } from './profile-search.service';
   <div class="searchBox">
   <h1>Profile Search</h1>
   <input #box (keyup.enter)="search(box.value)">
+  <ghv-app-messages></ghv-app-messages>
   </div>
   <ghv-profile-show></ghv-profile-show>
   </div>
