+++ src/app/profile-search/profile-search.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient} from '@angular/common/http';
 import { BehaviorSubject } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { tap, catchError } from 'rxjs/operators';
+import { HttpErrorHandler, HandleError } from '../http-error-handler.service';
 
 export interface ProfileInfo {
   login: string
@@ -18,9 +19,12 @@ export const searchUrl = 'https://api.github.com/users/';
 export class ProfileSearchService {
   profile: BehaviorSubject<ProfileInfo> = new BehaviorSubject({ login: '', name: '', avatar_url: '', repos_url: '' });
   repos: BehaviorSubject<any[]> = new BehaviorSubject([]);
+  private handleError: HandleError;
  
  
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient,
+    httpErrorHandler: HttpErrorHandler) {
+    this.handleError = httpErrorHandler.createHandleError('ProfileSearchService');
     this.profile.subscribe(({ repos_url }) => {
       if (repos_url) {
         // http request, set repoFetch to return value
@@ -33,7 +37,8 @@ export class ProfileSearchService {
 
   search(username: string) {
     this.http.get(searchUrl + username).pipe(
-      tap((result: ProfileInfo) => this.profile.next(result))
+      tap((result: ProfileInfo) => this.profile.next(result)),
+      catchError(this.handleError('Users', []))
     ).subscribe();
   }
